adding assignees to template structure

+++ notes/src/main/java/com/notes/notes/service/taskModuleServices/impl/TaskAssignmentServiceImpl.java
@@ -35,6 +35,11 @@ public class TaskAssignmentServiceImpl implements TaskAssignmentService {
                 taskAssignmentRepository.findByTaskAndRoleType(
                         task, TaskAssignment.AssignmentRole.MAIN_ASSIGNEE);
 
+        if (existingMain != null &&
+                existingMain.getUser().getUserId().equals(user.getUserId())) {
+            return existingMain;
+        }
+
         if (existingMain != null) {
             taskAssignmentRepository.delete(existingMain);
         }
